// src/application/dto/session/flow-step.enum.ts
export enum FlowStep {
  ENTRY = 'ENTRY',
  NON_RESIDENT_MENU = 'NON_RESIDENT_MENU',
  ASK_SUBUNIT_CODE = 'ASK_SUBUNIT_CODE',
  ASK_BUILDING_CODE = 'ASK_BUILDING_CODE',
  ASK_DEPARTMENT_NUMBER = 'ASK_DEPARTMENT_NUMBER',
  ADMIN_ASK_FIRST_NAME = 'ADMIN_ASK_FIRST_NAME',
  ADMIN_ASK_LAST_NAME = 'ADMIN_ASK_LAST_NAME',
  ADMIN_ASK_EMAIL = 'ADMIN_ASK_EMAIL',
  ADMIN_ASK_DOC_TYPE = 'ADMIN_ASK_DOC_TYPE',
  ADMIN_ASK_DOC_NUMBER = 'ADMIN_ASK_DOC_NUMBER',
  RESIDENT_MENU = 'RESIDENT_MENU',
  DONE = 'DONE',
}